# api/openapi/gateway/v1/gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "gateway/v1/gateway.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/idempotents": {
      "get": {
        "summary": "GetIdempotentToken",
        "operationId": "Gateway_GetIdempotentToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1IdempotentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/miners": {
      "get": {
        "summary": "ListMiner",
        "operationId": "Gateway_ListMiner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMinerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "description": "Offset is the starting point of the list for pagination.\n@gotags: form:\"offset\"",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "description": "Limit is the maximum number of miners to return.\n@gotags: form:\"limit\"",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "post": {
        "summary": "CreateMiner",
        "operationId": "Gateway_CreateMiner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appsv1beta1Miner"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Miner is the Schema for the miners API.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appsv1beta1Miner"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "put": {
        "summary": "UpdateMiner",
        "operationId": "Gateway_UpdateMiner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appsv1beta1Miner"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Miner is the Schema for the miners API.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appsv1beta1Miner"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/miners/{name}": {
      "get": {
        "summary": "GetMiner",
        "operationId": "Gateway_GetMiner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appsv1beta1Miner"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "name is the unique identifier of the miner to retrieve.\n@gotags: uri:\"name\"",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "delete": {
        "summary": "DeleteMiner",
        "operationId": "Gateway_DeleteMiner",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteMinerResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/minersets": {
      "get": {
        "summary": "ListMinerSet",
        "operationId": "Gateway_ListMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMinerSetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "description": "Offset is the starting point of the list for pagination.\n@gotags: form:\"offset\"",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "description": "Limit is the maximum number of minersets to return.\n@gotags: form:\"limit\"",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "post": {
        "summary": "CreateMinerSet",
        "operationId": "Gateway_CreateMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appsv1beta1MinerSet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "MinerSet ensures that a specified number of miners replicas are running at any given time.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appsv1beta1MinerSet"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "put": {
        "summary": "UpdateMinerSet",
        "operationId": "Gateway_UpdateMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appsv1beta1MinerSet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "MinerSet ensures that a specified number of miners replicas are running at any given time.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/appsv1beta1MinerSet"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/minersets/{name}": {
      "get": {
        "summary": "GetMinerSet",
        "operationId": "Gateway_GetMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/appsv1beta1MinerSet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "name is the unique identifier of the minerset to retrieve.\n@gotags: uri:\"name\"",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Gateway"
        ]
      },
      "delete": {
        "summary": "DeleteMinerSet",
        "operationId": "Gateway_DeleteMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteMinerSetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
    "/v1/minersets/{name}/scale": {
      "put": {
        "summary": "ScaleMinerSet",
        "operationId": "Gateway_ScaleMinerSet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ScaleMinerSetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GatewayScaleMinerSetBody"
            }
          }
        ],
        "tags": [
          "Gateway"
        ]
      }
    },
...
  },
  "definitions": {
...
    "appsv1beta1Miner": {
      "type": "object",
      "properties": {
        "metadata": {
          "$ref": "#/definitions/metav1ObjectMeta",
          "title": "Standard object's metadata.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata\n+optional"
        },
        "spec": {
          "$ref": "#/definitions/v1beta1MinerSpec",
          "title": "Specification of the desired behavior of the miner.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional"
        },
        "status": {
          "$ref": "#/definitions/v1beta1MinerStatus",
          "title": "Most recently observed status of the miner.\nThis data may not be up to date.\nPopulated by the system.\nRead-only.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional"
        }
      },
      "description": "Miner is the Schema for the miners API."
    },
    "appsv1beta1MinerSet": {
      "type": "object",
      "properties": {
        "metadata": {
          "$ref": "#/definitions/metav1ObjectMeta",
          "title": "If the Labels of a MinerSet are empty, they are defaulted to\nbe the same as the Miner(s) that the MinerSet manages.\nStandard object's metadata.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata\n+optional"
        },
        "spec": {
          "$ref": "#/definitions/v1beta1MinerSetSpec",
          "title": "Spec defines the specification of the desired behavior of the MinerSet.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional"
        },
        "status": {
          "$ref": "#/definitions/v1beta1MinerSetStatus",
          "title": "Status is the most recently observed status of the MinerSet.\nThis data may be out of date by some window of time.\nPopulated by the system.\nRead-only.\nMore info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status\n+optional"
        }
      },
      "description": "MinerSet ensures that a specified number of miners replicas are running at any given time."
    },
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1DeleteMinerResponse": {
      "type": "object",
      "description": "DeleteMinerResponse represents the response message for a successful miner deletion."
    },
...
    "v1IdempotentResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
...
    "v1ListMinerResponse": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64",
          "description": "TotalCount is the total number of miners matching the query."
        },
        "miners": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/gatewayv1Miner"
          },
          "description": "Miner is the list of miners in the current page."
        }
      },
      "description": "ListMinerResponse represents the response message for listing miners."
    },
    "v1ListMinerSetResponse": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64",
          "description": "TotalCount is the total number of minersets matching the query."
        },
        "minerSets": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/gatewayv1MinerSet"
          },
          "description": "MinerSet is the list of minersets in the current page."
        }
      },
      "description": "ListMinerSetResponse represents the response message for listing minersets."
    },
...
    "v1ScaleMinerSetResponse": {
      "type": "object"
    },
...
  }
}